separated tour stuff into three separate files
